// web/src/api/generated/models/authSignInResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * POS Cafe API
 * Backend API for Cafe Point of Sale System built with Vertical Slice Architecture and Idiomatic Go.
 * OpenAPI spec version: 1.0
 */
import type { AuthStaffProfileResponse } from './authStaffProfileResponse';

export interface AuthSignInResponse {
  staff?: AuthStaffProfileResponse;
  token?: string;
}